% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim_data.R
\name{gen_sim_data}
\alias{gen_sim_data}
\title{Summarize empirical power based on an input dataset of simulated experiments. Will return a list of results: plot of power curves (powerplot) and data frame of power results (powerresults)}
\usage{
gen_sim_data(
  n_sample_size = 2000,
  n_sims = 2000,
  est_sensitivity = 0.84,
  est_specificity = 0.85,
  sens_lower_boundary = 0.8,
  spec_lower_boundary = 0.8,
  ppv_lower_boundary = 0.4,
  npv_lower_boundary = 0.9,
  p_vec = c(0.25, 0.1, 0.05, 0.05, 0.1, 0.02, 0.01, 0.07, 0.35),
  pos_group = c(1, 2, 3, 7),
  neg_group = c(4, 5, 6, 9),
  alpha = 0.05,
  target_power = 0.9
)
}
\arguments{
\item{n_sample_size}{Sample size for the overall study}

\item{n_sims}{Number of simulation replicates to run (default = 2000)}

\item{est_sensitivity}{Estimate (under alternative) of the sensitivity of the test}

\item{est_specificity}{Estimate (under alternative) of the specificity of the test}

\item{sens_lower_boundary}{Lower limit for the Exact confidence interval}

\item{spec_lower_boundary}{Lower limit for the Exact confidence interval}

\item{ppv_lower_boundary}{Lower limit for the Exact confidence interval}

\item{npv_lower_boundary}{Lower limit for the Exact confidence interval}

\item{p_vec}{The multinomial vector of probabilities across 9 possible bins for participants. Vector entries may be zero but must sum to 1.  Using the pos_group and neg_group, these categories can be groupped into higher order positive and negative groups}

\item{pos_group}{A set of indices (1 -9 as integers) for the classes that are considered disease positive}

\item{neg_group}{A set of indices (1 - 9 as integers) for the classes that are considered negative for the disease. Any indices in the 1 to 9 range not in the pos or neg group will be considered indeterminate}

\item{alpha}{The level of significance for the tests.  Defaults to 0.05.  Decrease to allow for multiple testing (e.g., 0.0125 for tests of Sens, Spec, PPV, and NPV)}
}
\description{
Summarize empirical power based on an input dataset of simulated experiments. Will return a list of results: plot of power curves (powerplot) and data frame of power results (powerresults)
}
