% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_power.R
\name{summarize_power}
\alias{summarize_power}
\title{Summarize empirical power based on an input dataset of simulated experiments. Will return a list of results: plot of power curves (powerplot) and data frame of power results (powerresults)}
\usage{
summarize_power(indf)
}
\arguments{
\item{indf}{A data frame of simulated data generated from the gen_sim_data function}
}
\description{
Summarize empirical power based on an input dataset of simulated experiments. Will return a list of results: plot of power curves (powerplot) and data frame of power results (powerresults)
}
